apps/api/mentors/requirements.txt
blinker==1.6.2
click==8.1.7
et-xmlfile==1.1.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.7
